requirements.txt
wordcloud
pillow
numpy
cachier
nltk
